backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
urllib3<2.0.0
typing-extensions==4.8.0
yt-dlp==2025.8.22
pydub==0.25.1
openai-whisper==20231117
ffmpeg-python==0.2.0
sqlalchemy==2.0.23
librosa==0.10.1
scipy==1.11.4
numpy==1.24.3
# Remove OpenAI library dependency since we're using requests directly
# openai>=1.12.0,<2.0.0
python-dotenv==1.0.0
